Genome merge/split integration in a single utils pipeline
